keep `rawTypes` and `types` independent in `JavaClass.Interfaces`
By mapping the `types` to `rawTypes` we would hide any wrong assumptions that this should always be in sync. I think we should rather find out, if this is the case. Also, meanwhile I found cases, where the `rawTypes` and `types` don't even match 1-2-1 (e.g. generically typed private inner class constructor). So I think it is better to keep exactly what we import from the bytecode and analyze it, if we run into strange cases.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>